Moved functions from utils controller to gen_main generator
