+++ goagen/gen_controller/doc.go
@@ -1,9 +1,6 @@
 /*
-Package gencontroller provides a generator for a skeleton goa application.
-This generator generates the code for a basic "controller" package and is mainly intended as a way to
-create the controllers of new applications.
-The generator creates a controller.go file and one file per resource listed in the API metadata.
-If a file already exists it skips its creation unless the flag --force is provided on the command
-line in which case it overrides the content of existing files.
+Package gencontroller generates the controller code for a given design resource.
+This generator is intended for use when resources are added to the design
+after the initial bootstrapping.
 */
 package gencontroller
